+++ .npmrc
@@ -1,2 +1,6 @@
+; Enforces that the user is `npm install`ing with the correct node version.
 engine-strict=true
+
+; Workaround for conflict between the default location(s) of node-forge and the
+; location expected by Typescript, Jasmine, and Electron.
 prefer-dedupe=true
