+++ src/server_manager/electron_app/fetch.ts
@@ -18,7 +18,7 @@ import {urlToHttpOptions} from 'url';
 
 import type {IncomingMessage} from 'http';
 
-import type {HttpRequest, HttpResponse} from './types';
+import type {HttpRequest, HttpResponse} from '../infrastructure/path_api';
 
 export const fetchWithPin = async (
   req: HttpRequest,
