+++ src/server_manager/electron_app/preload.ts
@@ -18,7 +18,7 @@ import {URL} from 'url';
 
 import * as digitalocean_oauth from './digitalocean_oauth';
 import * as gcp_oauth from './gcp_oauth';
-import {HttpRequest, HttpResponse} from './http/types';
+import {HttpRequest, HttpResponse} from '../infrastructure/path_api';
 import {redactManagerUrl} from './util';
 
 // This file is run in the renderer process *before* nodeIntegration is disabled.
