+++ src/server_manager/types/preload.d.ts
@@ -14,8 +14,8 @@
 
 // Functions made available to the renderer process via preload.ts.
 
-type HttpRequest = import('../electron_app/http/types').HttpRequest;
-type HttpResponse = import('../electron_app/http/types').HttpResponse;
+type HttpRequest = import('../infrastructure/path_api').HttpRequest;
+type HttpResponse = import('../infrastructure/path_api').HttpResponse;
 
 declare function fetchWithPin(request: HttpRequest, fingerprint: string): Promise<HttpResponse>;
 declare function openImage(basename: string): void;
