+++ src/server_manager/web_app/manual_server.ts
@@ -14,7 +14,7 @@
 
 import {hexToString} from '../infrastructure/hex_encoding';
 import * as server from '../model/server';
-import {makePathApiClient} from './path_api';
+import {makePathApiClient} from './fetcher';
 
 import {ShadowboxServer} from './shadowbox_server';
