+++ src/server_manager/web_app/shadowbox_server.ts
@@ -15,7 +15,7 @@
 import * as semver from 'semver';
 
 import * as server from '../model/server';
-import {PathApiClient} from './path_api';
+import {PathApiClient} from '../infrastructure/path_api';
 
 interface AccessKeyJson {
   id: string;
